+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed `$distinct()` to not return missing values per default.
 * Added `na_rm` argument to `$distinct()`.
+* Renamed `as_sqlite()` to `as_sqlite_backend()`
 
 # mlr3db 0.1.1